COMMS/PYTHON/requirements.txt
enum34==1.1.6
pyserial==3.4
robust-serial==0.1
